requirements.txt
beautifulsoup4
requests
neo4j
python-dotenv
openai>=1.51.0
agno
